Binh

+++ admin/controllers/danhmucController.php
@@ -16,6 +16,13 @@ class danhmucController
     {
         if ($_GET['id']) {
             $id = $_GET['id'];
+            if ($this->danhmucModel->checkDanhmucHasProducts($id)) {
+                // Nếu có sản phẩm, hiển thị thông báo lỗi
+                $errorMessage = "Không thể xóa danh mục vì danh mục này đang chứa sản phẩm!";
+                $listDanhmuc = $this->danhmucModel->getAllDanhmuc();
+                require_once 'views/danhmuc/danhmuc.php';
+                return;
+            }
             $id = $this->danhmucModel->deleteDanhmuc($id);
         }
         header('location:index.php?act=danhmuc');
@@ -42,10 +49,15 @@ class danhmucController
         if ($id) {
             $iddm = $this->danhmucModel->getIdDanhmuc($id);
             if (isset($_POST['update_danhmuc'])) {
-                $name = $_POST['name'];
-                $this->danhmucModel->updateDanhmuc($id, $name);
-                header('location:index.php?act=danhmuc');
-                exit;
+                $name = trim($_POST['name']);
+                if($name==""){
+                    $error="Không được để trống tên danh mục";
+                }else{
+                    $this->danhmucModel->updateDanhmuc($id, $name);
+                    header('location:index.php?act=danhmuc');
+                    exit;
+                }
+               
             }
         }
         require_once 'views/danhmuc/editDanhmuc.php';
